chore: add src/utils/date.js
